# laboratorul9/security/main/java/com/example/laboratorul8/model/Evaluation.java
package com.example.laboratorul8.model;

import com.fasterxml.jackson.annotation.JsonManagedReference;
import jakarta.json.bind.annotation.JsonbTransient;
import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.hibernate.annotations.CreationTimestamp;

import java.time.LocalDateTime;
import java.time.format.DateTimeFormatter;

@Entity
@Data
@AllArgsConstructor
@NoArgsConstructor
@NamedQueries({
        @NamedQuery(
                name = "Evaluation.findByTeacher",
                query = "SELECT e FROM Evaluation e WHERE e.teacher = :teacher"
        ),
        @NamedQuery(name = "Evaluation.getTeacherStatistics",
                query = "SELECT t.name, AVG(e.grade), COUNT(e.id) FROM Evaluation e " +
                        "JOIN e.teacher t " +
                        "GROUP BY t.name")
})
public class Evaluation {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @ManyToOne(fetch = FetchType.EAGER)
    @JoinColumn(name = "student_id", nullable = false)
    private Student student;

    @ManyToOne(fetch = FetchType.EAGER)
    @JoinColumn(name = "teacher_id", nullable = false)
    private Teacher teacher;

    @ManyToOne(fetch = FetchType.EAGER)
    @JoinColumn(name = "course_id", nullable = false)
    @JsonManagedReference
    private Course course;

    private String activity;

    private int grade;
    private String comment;

    @CreationTimestamp
    @Column(updatable = false)
    private LocalDateTime createdAt;

    private String registrationNumber;

    @JsonbTransient
    public LocalDateTime getCreatedAtOriginal() {
        return createdAt;
    }

    public String getCreatedAt() {
        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm");
        return createdAt.format(formatter);
    }

}